Move Kafka Streams classes to dedicated `streams` package
