+++ src/main/java/org/dependencytrack/event/kafka/streams/processor/VulnerabilityScanResultProcessor.java
@@ -1,4 +1,4 @@
-package org.dependencytrack.event.kafka.processor;
+package org.dependencytrack.event.kafka.streams.processor;
 
 import alpine.Config;
 import alpine.common.logging.Logger;
@@ -625,7 +625,7 @@ public class VulnerabilityScanResultProcessor extends ContextualFixedKeyProcesso
                         .setLevel(LEVEL_INFORMATIONAL)
                         .setTimestamp(Timestamps.now())
                         .setTitle(generateTitle(policyAnalysis.getState(), policyAnalysis.getSuppressed(), analysisStateChange, suppressionChange))
-                        .setContent(generateNotificationContent(policyAnalysis))
+                        .setContent("An analysis decision was made to a finding affecting a project")
                         .setSubject(Any.pack(subject))
                         .build())
                 .orElse(null);
