+++ src/main/webapp/WEB-INF/web.xml
@@ -48,7 +48,7 @@
         <listener-class>org.dependencytrack.event.EventSubsystemInitializer</listener-class>
     </listener>
     <listener>
-        <listener-class>org.dependencytrack.event.kafka.KafkaStreamsInitializer</listener-class>
+        <listener-class>org.dependencytrack.event.kafka.streams.KafkaStreamsInitializer</listener-class>
     </listener>
     <listener>
         <listener-class>org.dependencytrack.event.PurlMigrator</listener-class>
