+++ src/test/java/org/dependencytrack/event/kafka/streams/KafkaStreamsTest.java
@@ -1,13 +1,11 @@
-package org.dependencytrack.event.kafka;
+package org.dependencytrack.event.kafka.streams;
 
 import net.mguenther.kafka.junit.ExternalKafkaCluster;
 import net.mguenther.kafka.junit.TopicConfig;
 import org.apache.kafka.streams.KafkaStreams;
 import org.apache.kafka.streams.StreamsConfig;
-import org.apache.kafka.streams.Topology;
-import org.dependencytrack.AbstractPostgresEnabledTest;
-import org.dependencytrack.event.kafka.serialization.KafkaProtobufDeserializer;
-import org.dependencytrack.proto.notification.v1.Notification;
+import org.dependencytrack.PersistenceCapableTest;
+import org.dependencytrack.event.kafka.KafkaTopics;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -17,11 +15,10 @@ import org.testcontainers.utility.DockerImageName;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
-import java.util.function.Supplier;
 
 import static org.dependencytrack.assertion.Assertions.assertConditionWithTimeout;
 
-abstract class KafkaStreamsPostgresTest extends AbstractPostgresEnabledTest {
+abstract class KafkaStreamsTest extends PersistenceCapableTest {
 
     @Rule
     public RedpandaContainer container = new RedpandaContainer(DockerImageName
@@ -29,20 +26,10 @@ abstract class KafkaStreamsPostgresTest extends AbstractPostgresEnabledTest {
 
     KafkaStreams kafkaStreams;
     ExternalKafkaCluster kafka;
-    private final Supplier<Topology> topologySupplier;
     private Path kafkaStreamsStateDirectory;
-
-    protected KafkaStreamsPostgresTest() {
-        this(new KafkaStreamsTopologyFactory()::createTopology);
-    }
-
-    protected KafkaStreamsPostgresTest(final Supplier<Topology> topologySupplier) {
-        this.topologySupplier = topologySupplier;
-    }
 
     @Before
     public void setUp() throws Exception {
-        super.setUp();
         kafka = ExternalKafkaCluster.at(container.getBootstrapServers());
 
         kafka.createTopic(TopicConfig
@@ -70,7 +57,7 @@ abstract class KafkaStreamsPostgresTest extends AbstractPostgresEnabledTest {
         streamsConfig.put(StreamsConfig.STATE_DIR_CONFIG, kafkaStreamsStateDirectory.toString());
         streamsConfig.put(StreamsConfig.NUM_STREAM_THREADS_CONFIG, "3");
 
-        kafkaStreams = new KafkaStreams(topologySupplier.get(), streamsConfig);
+        kafkaStreams = new KafkaStreams(new KafkaStreamsTopologyFactory().createTopology(), streamsConfig);
         kafkaStreams.start();
 
         assertConditionWithTimeout(() -> KafkaStreams.State.RUNNING == kafkaStreams.state(), Duration.ofSeconds(5));
@@ -84,15 +71,6 @@ abstract class KafkaStreamsPostgresTest extends AbstractPostgresEnabledTest {
         if (kafkaStreamsStateDirectory != null) {
             kafkaStreamsStateDirectory.toFile().delete();
         }
-        super.tearDown();
-    }
-
-    public static class NotificationDeserializer extends KafkaProtobufDeserializer<Notification> {
-
-        public NotificationDeserializer() {
-            super(Notification.parser());
-        }
-
     }
 
 }
